reformat and moving build_storage** to views
